model/requirements.txt
numpy
pandas
matplotlib
scikit-learn
torch
torchvision
tqdm
transformers
datasets
huggingface_hub
python-dotenv
graphviz
bitsandbytes
accelerate
peft
Pillow
networkx
pydot
openai
psycopg2-binary
pgvector
fastapi
uvicorn
pydantic
python-multipart
scikit-image
lpips
azure-identity
azure-ai-projects
